WIP remove pre-loading and caching of all items
