+++ api/lib/project-item-node-to-github-project-item.js
@@ -6,7 +6,7 @@ import { itemFieldsNodesToFieldsMap } from "./item-fields-nodes-to-fields-map.js
  * Takes a GraphQL `ProjectNextItem` node and returns a `ProjectItem` object
  * in the format we return it from the GitHubProject API.
  *
- * @param {import("../..").GitHubProjectStateWithFields | import("../..").GitHubProjectStateWithItems} state
+ * @param {import("../..").GitHubProjectStateWithFields | import("../..").GitHubProjectStateWithFields} state
  * @param {any} itemNode
  * @returns {import("../..").GitHubProjectItem}
  */
